make gcc with older headers happy /usr/include/c++/4.4/bits/basic_string.h only has  to_string(long long __val)  to_string(unsigned long long __val)  to_string(long double __val)

+++ ext/hmsearch/hmsearch.cc
@@ -324,7 +324,7 @@ bool HmSearchImpl::insert(const hash_string& hash,
         std::stringstream s;
         s << "INSERT INTO partition" << i
         << "  VALUES ($1, $2)";
-        _db->prepare("insert_"+std::to_string(i), s.str());
+        _db->prepare("insert_"+std::to_string(static_cast<long long>(i)), s.str());
     }
     pqxx::work W(*_db);
     for (int i = 0; i < _partitions; i++) {
@@ -335,7 +335,7 @@ bool HmSearchImpl::insert(const hash_string& hash,
         pqxx::binarystring key_blob(key, _partition_bytes);
         pqxx::binarystring hash_blob(hash.data(), hash.length());
         
-        W.prepared("insert_"+std::to_string(i))(hash_blob)(key_blob).exec();
+        W.prepared("insert_"+std::to_string(static_cast<long long>(i)))(hash_blob)(key_blob).exec();
     }
     
     W.commit();
@@ -416,7 +416,7 @@ HmSearchImpl::hash_string HmSearchImpl::get_multiple_keys(
     
     pqxx::nontransaction n(*_db);
     pqxx::binarystring key_blob(key, _partition_bytes);
-    pqxx::result res = n.prepared("select_"+std::to_string(partition))(key_blob).exec();
+    pqxx::result res = n.prepared("select_"+std::to_string(static_cast<long long>(partition)))(key_blob).exec();
     
     for (pqxx::result::const_iterator c = res.begin(); c != res.end(); ++c) {
         pqxx::binarystring hash_result(c[0]);
@@ -456,11 +456,11 @@ void HmSearchImpl::get_candidates(
         << ".key = x.key";
         std::string sql;
         sql.append(s.str());
-        _db->prepare("select_multiple_"+std::to_string(i), sql);
+        _db->prepare("select_multiple_"+std::to_string(static_cast<long long>(i)), sql);
         
         sql.clear();
         sql.append(single.str());
-        _db->prepare("select_"+std::to_string(i), sql);
+        _db->prepare("select_"+std::to_string(static_cast<long long>(i)), sql);
     }
     for (int i = 0; i < _partitions; i++) {
         hash_string hashes;
@@ -477,7 +477,7 @@ void HmSearchImpl::get_candidates(
         
         // Get 1-variant matches
         pqxx::nontransaction n(*_db);
-        pqxx::prepare::invocation prep = n.prepared("select_multiple_"+std::to_string(i));
+        pqxx::prepare::invocation prep = n.prepared("select_multiple_"+std::to_string(static_cast<long long>(i)));
         int pbyte = (i * _partition_bits) / 8;
         int count = 0;
         for (int pbit = i * _partition_bits; bits > 0; pbit++, bits--, count++) {
